move tutorial files to separate folder
