+++ src/assets/tutorial/CNN.png
