+++ src/assets/tutorial/FOX.png
